Redirect 2.3.2 * Use class alias for CSRF validation
