+++ Redirect/SID/InstantEFT/etc/module.xml
@@ -9,7 +9,7 @@
  */
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
-    <module name="SID_InstantEFT" setup_version="2.3.1">
+    <module name="SID_InstantEFT" setup_version="2.3.2">
         <sequence>
             <module name="Magento_Checkout"/>
             <module name="Magento_Sales"/>
